+++ includes/class-plugin.php
@@ -24,6 +24,7 @@ final class WCCR_Plugin {
 		$settings_repository  = new WCCR_Settings_Repository();
 		$cart_repository      = new WCCR_Cart_Repository();
 		$email_log_repository = new WCCR_Email_Log_Repository();
+		$stats_repository     = new WCCR_Stats_Repository();
 		$locale_resolver      = new WCCR_Locale_Resolver_Manager();
 		$email_eligibility    = new WCCR_Email_Eligibility_Service( $email_log_repository );
 		$cart_capture_service = new WCCR_Cart_Capture_Service( $cart_repository, $locale_resolver );
@@ -33,8 +34,8 @@ final class WCCR_Plugin {
 		$email_renderer       = new WCCR_Email_Renderer( $coupon_service );
 		$email_scheduler      = new WCCR_Email_Scheduler( $cart_repository, $email_log_repository, $settings_repository, $email_eligibility, $coupon_service, $email_renderer, $recovery_service );
 		$detector             = new WCCR_Abandoned_Cart_Detector( $cart_repository, $settings_repository );
-		$cleanup_service      = new WCCR_Cleanup_Service( $cart_repository, $email_log_repository, $settings_repository );
-		$stats_service        = new WCCR_Stats_Service( $cart_repository, $email_log_repository );
+		$cleanup_service      = new WCCR_Cleanup_Service( $cart_repository, $email_log_repository, $stats_repository, $settings_repository );
+		$stats_service        = new WCCR_Stats_Service( $cart_repository, $email_log_repository, $stats_repository );
 
 		$checkout = new WCCR_Checkout_Capture_Coordinator(
 			new WCCR_Classic_Checkout_Capture_Adapter( $cart_capture_service ),
@@ -47,7 +48,7 @@ final class WCCR_Plugin {
 
 		$admin = new WCCR_Admin_Menu(
 			$settings_page,
-			new WCCR_Abandoned_Carts_Page( $cart_repository, $email_log_repository, $settings_repository, $email_eligibility, $stats_service ),
+			new WCCR_Abandoned_Carts_Page( $cart_repository, $email_log_repository, $stats_repository, $settings_repository, $email_eligibility, $stats_service ),
 			new WCCR_Stats_Page( $stats_service )
 		);
 		$admin->register_hooks();